requirements.txt
flask 
flask-restful 
flask-sqlalchemy 
flask-migrate 
flask-jwt-extended 
flask-cors 
marshmallow 
python-dotenv 
psycopg2-binary